* point light

+++ include/Graphic/Shader.hpp
@@ -33,7 +33,7 @@ class Shader {
 
     void setInt(const std::string& name, int value) const;
 
-    void setFloat(const std::string& name, int value) const;
+    void setFloat(const std::string& name, float value) const;
 
     void setVec3(const std::string& name, const glm::vec3& value) const;
